requirements.txt
flask
werkzeug
langchain
langchain-openai
langchain-community
openai
python-dotenv
boto3
chromadb
pypdf
tiktoken
unstructured